Split QgsLabelingGui off into QgsTextFormatWidget
New widget allows for setting just the formatting properties
of text
