+++ src/app/qgslabelinggui.h
@@ -15,26 +15,18 @@
  *                                                                         *
  ***************************************************************************/
 
-#ifndef QgsLabelingGUI_H
-#define QgsLabelingGUI_H
+#ifndef QGSLABELINGGUI_H
+#define QGSLABELINGGUI_H
 
-#include <QDialog>
-#include <QFontDatabase>
-#include <ui_qgslabelingguibase.h>
-#include "qgsstringutils.h"
 #include "qgspallabeling.h"
+#include "qgstextformatwidget.h"
 
-class QgsVectorLayer;
-class QgsMapCanvas;
-class QgsCharacterSelectorDialog;
-
-class APP_EXPORT QgsLabelingGui : public QWidget, private Ui::QgsLabelingGuiBase, private QgsExpressionContextGenerator
+class APP_EXPORT QgsLabelingGui : public QgsTextFormatWidget, private QgsExpressionContextGenerator
 {
     Q_OBJECT
 
   public:
     QgsLabelingGui( QgsVectorLayer* layer, QgsMapCanvas* mapCanvas, const QgsPalLayerSettings* settings, QWidget* parent );
-    ~QgsLabelingGui();
 
     QgsPalLayerSettings layerSettings();
     void writeSettingsToLayer();
@@ -48,111 +40,30 @@ class APP_EXPORT QgsLabelingGui : public QWidget, private Ui::QgsLabelingGuiBase
 
     void setLabelMode( LabelMode mode );
 
-  signals:
-    void widgetChanged();
+    void setLayer( QgsMapLayer* layer );
 
   public slots:
-    void setLayer( QgsMapLayer* layer );
-    void setDockMode( bool enabled );
-    void connectValueChanged( QList<QWidget*> widgets, const char* slot );
 
-    void init();
-    void collapseSample( bool collapse );
     void apply();
-    void changeTextColor( const QColor &color );
-    void changeBufferColor( const QColor &color );
 
     void updateUi();
-    void updatePreview();
-    void scrollPreview();
-    void updatePlacementWidgets();
-    void updateSvgWidgets( const QString& svgPath );
-
-    void on_mFontSizeSpinBox_valueChanged( double d );
-    void on_mFontCapitalsComboBox_currentIndexChanged( int index );
-    void on_mFontFamilyCmbBx_currentFontChanged( const QFont& f );
-    void on_mFontStyleComboBox_currentIndexChanged( const QString & text );
-    void on_mFontUnderlineBtn_toggled( bool ckd );
-    void on_mFontStrikethroughBtn_toggled( bool ckd );
-    void on_mFontWordSpacingSpinBox_valueChanged( double spacing );
-    void on_mFontLetterSpacingSpinBox_valueChanged( double spacing );
-    void on_mFontSizeUnitWidget_changed();
-    void on_mFontMinPixelSpinBox_valueChanged( int px );
-    void on_mFontMaxPixelSpinBox_valueChanged( int px );
-    void on_mBufferUnitWidget_changed();
-    void on_mCoordXDDBtn_dataDefinedActivated( bool active );
-    void on_mCoordYDDBtn_dataDefinedActivated( bool active );
-
-    void on_mShapeTypeCmbBx_currentIndexChanged( int index );
-    void on_mShapeRotationCmbBx_currentIndexChanged( int index );
-    void on_mShapeSVGParamsBtn_clicked();
-    void on_mShapeSVGSelectorBtn_clicked();
-
-    void on_mPreviewTextEdit_textChanged( const QString & text );
-    void on_mPreviewTextBtn_clicked();
-    void on_mPreviewBackgroundBtn_colorChanged( const QColor &color );
-    void on_mDirectSymbLeftToolBtn_clicked();
-    void on_mDirectSymbRightToolBtn_clicked();
-    void on_mChkNoObstacle_toggled( bool active );
-    void on_chkLineOrientationDependent_toggled( bool active );
-
-    void on_mToolButtonConfigureSubstitutes_clicked();
 
   protected:
     void blockInitSignals( bool block );
-    void blockFontChangeSignals( bool blk );
-    void setPreviewBackground( const QColor& color );
     void syncDefinedCheckboxFrame( QgsDataDefinedButton* ddBtn, QCheckBox* chkBx, QFrame* f );
-    void populateFontCapitalsComboBox();
-    void populateFontStyleComboBox();
-    void populatePlacementMethods();
-    void populateFieldNames();
     void populateDataDefinedButtons( QgsPalLayerSettings& s );
     /** Sets data defined property attribute to map */
     void setDataDefinedProperty( const QgsDataDefinedButton* ddBtn, QgsPalLayerSettings::DataDefinedProperties p, QgsPalLayerSettings& lyr );
-    void updateFont( const QFont& font );
 
   private:
     QgsVectorLayer* mLayer;
-    QgsMapCanvas* mMapCanvas;
     const QgsPalLayerSettings* mSettings;
     LabelMode mMode;
-    QFontDatabase mFontDB;
-    QgsCharacterSelectorDialog* mCharDlg;
-
-    QButtonGroup* mQuadrantBtnGrp;
-    QButtonGroup* mDirectSymbBtnGrp;
-    QButtonGroup* mUpsidedownBtnGrp;
-
-    QButtonGroup* mPlacePointBtnGrp;
-    QButtonGroup* mPlaceLineBtnGrp;
-    QButtonGroup* mPlacePolygonBtnGrp;
-
-    // background reference font
-    QFont mRefFont;
-    bool mDockMode;
-    int mPreviewSize;
-
-    int mMinPixelLimit;
-
-    bool mLoadSvgParams;
-
-    QgsStringReplacementCollection mSubstitutions;
-
-    void enableDataDefinedAlignment( bool enable );
 
     QgsExpressionContext createExpressionContext() const override;
 
-  private slots:
-    void optionsStackedWidget_CurrentChanged( int indx );
-    void showBackgroundRadius( bool show );
-    void showBackgroundPenStyle( bool show );
-    void on_mShapeSVGPathLineEdit_textChanged( const QString& text );
-    void updateLinePlacementOptions();
-    void onSubstitutionsChanged( const QgsStringReplacementCollection& substitutions );
-    void previewScaleChanged( double scale );
 };
 
-#endif
+#endif // QGSLABELINGGUI_H
