+++ src/gui/qgssubstitutionlistwidget.cpp
